requirements.txt
certifi==2020.12.5
chardet==4.0.0
idna==2.10
PyJWT==2.0.1
PyYAML==5.4.1
requests==2.25.1
urllib3==1.26.4
